Переделываю функции
переделал функцию сборки. Осталось решить проблему с PAYLOAD_SIZE. Не успел доделать функцию разбора пакетов. Дальше нужно будет перепроверять все остальные шаги.

+++ func.c
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <string.h>
 
-void func_5(uint32_t* buf)
+void func_5(uint8_t* buf)
 {
 	for (uint32_t i = 0; i < 10; i++)
 	{
